+++ core.js
@@ -1,4 +1,4 @@
-/*jshint -W117, -W030, -W016 */
+/*jshint -W117, -W030, -W016, -W084 */
 /* global
  Velocity:true,
  DEBUG:true
@@ -20,7 +20,7 @@ DEBUG = !!process.env.VELOCITY_DEBUG;
  *
  * @class Velocity
  */
-Velocity = {};
+Velocity = Velocity || {};
 
 (function () {
   'use strict';
@@ -35,51 +35,31 @@ Velocity = {};
     DEBUG && console.log('[velocity] ' + (process.env.IS_MIRROR ? 'Mirror detected - ' : '') + 'Not adding velocity core');
     return;
   }
-
-  var getAssetPath = function (packageName, fileName) {
-    var serverAssetsPath = path.join(
-      findAppDir(), '.meteor', 'local', 'build', 'programs', 'server', 'assets'
-    );
-
-    packageName = packageName.replace(':', '_');
-
-    return path.join(serverAssetsPath, 'packages', packageName, fileName);
-  };
-
-  var escapeShellArgument = function (argument) {
-    if (!/(["'`\\ ])/.test(argument)) {
-      return argument;
-    } else {
-      return '"' + argument.replace(/(["'`\\])/g, '\\$1') + '"';
-    }
-  }
+  DEBUG && console.log('[velocity] adding velocity core');
 
   var _ = Npm.require('lodash'),
       fs = Npm.require('fs'),
-      fse = Npm.require('fs-extra'),
-      outputFile = Meteor.wrapAsync(fse.outputFile),
-      copyFile = Meteor.wrapAsync(fse.copy),
       path = Npm.require('path'),
-      url = Npm.require('url'),
-      Rsync = Npm.require('rsync'),
       child_process = Npm.require('child_process'),
       chokidar = Npm.require('chokidar'),
       mkdirp = Npm.require('mkdirp'),
       _config = {},
-      _preProcessors = [],
-      _postProcessors = [],
       _watcher,
-      FIXTURE_REG_EXP = new RegExp('-fixture.(js|coffee)$'),
-      DEFAULT_FIXTURE_PATH = getAssetPath('velocity:core', 'default-fixture.js'),
-      MIRROR_PID_VAR_TEMPLATE = 'mirror.@PORT.pid';
+      _velocityStarted = false,
+      _velocityStartupFunctions = [],
+      FIXTURE_REG_EXP = new RegExp('-fixture.(js|coffee)$');
+
 
   Meteor.startup(function initializeVelocity () {
     DEBUG && console.log('[velocity] Server startup');
     DEBUG && console.log('[velocity] app dir', Velocity.getAppPath());
-    DEBUG && console.log('velocity config =', JSON.stringify(_config, null, 2));
+    DEBUG && console.log('[velocity] config =', JSON.stringify(_config, null, 2));
 
-    // kick-off everything
+    //kick-off everything
     _reset(_config);
+
+    _initWatcher(_config, _triggerVelocityStartupFunctions);
+
   });
 
 //////////////////////////////////////////////////////////////////////
@@ -87,6 +67,23 @@ Velocity = {};
 //
 
   _.extend(Velocity, {
+
+    /**
+     * Run code when Velocity is started. Velocity is considered started when the file watcher has
+     * completed the scan of the  file system
+     *
+     * @method startup
+     * @return {function} A function to run on startup
+     */
+    startup: function (func) {
+      if (_velocityStarted) {
+        DEBUG && console.log('[velocity] Velocity already started. Immediately calling func');
+        func();
+      } else {
+        DEBUG && console.log('[velocity] Velocity not started. Queueing func');
+        _velocityStartupFunctions.push(func);
+      }
+    },
 
     /**
      * Get application directory path.
@@ -98,15 +95,6 @@ Velocity = {};
       return findAppDir();
     },
 
-    /**
-     * Get path to directory of the 'mirror' application.
-     *
-     * @method getMirrorPath
-     * @return {String} mirror directory path
-     */
-    getMirrorPath: function () {
-      return path.join(Velocity.getAppPath(), '.meteor', 'local', '.mirror');
-    },
 
     /**
      * Get path to application's 'tests' directory
@@ -119,26 +107,15 @@ Velocity = {};
     },
 
     /**
-     * Add a callback which will execute when the mirror application
-     * is sync'ed up with latest application code.  Preprocessors will
-     * execute before fixtures are added.
-     *
-     * @method addPreProcessor
-     * @param {Function} preprocessor
-     */
-    addPreProcessor: function (preProcessor) {
-      _preProcessors.push(preProcessor);
-    },
-
-    /**
      * Add a callback which will execute after all tests have completed
      * and after the aggregate test results have been reported.
      *
      * @method addPostProcessor
-     * @param {Function} reporter
+     * @param {Function} processor
      */
-    addPostProcessor: function (reporter) {
-      _postProcessors.push(reporter);
+    postProcessors: [],
+    addPostProcessor: function (processor) {
+      Velocity.postProcessors.push(processor);
     },
 
     /**
@@ -150,39 +127,38 @@ Velocity = {};
      */
     getReportGithubIssueMessage: function () {
       return 'Please report the issue here: ' +
-             'https://github.com/meteor-velocity/velocity/issues';
+        'https://github.com/meteor-velocity/velocity/issues';
+    },
+
+    /**
+     * Registers a testing framework plugin.
+     *
+     * @method registerTestingFramework
+     * @param {String} name                       The name of the testing framework.
+     * @param {Object} [options]                  Options for the testing framework.
+     * @param {String} options.disableAutoReset   Velocity's reset cycle will skip reports and logs for this framework
+     *                                            It will be the responsibility of the framework to clean up its ****!
+     * @param {String} options.regex              The regular expression for test files that should be assigned
+     *                                            to the testing framework.
+     *                                            The path relative to the tests
+     *                                            folder is matched against it.
+     *                                            The default is "name/.+\.js$"
+     *                                            (name is the testing framework name).
+     * @param options.sampleTestGenerator {Function} sampleTestGenerator
+     *    returns an array of fileObjects with the following fields:
+     * @param options.sampleTestGenerator.path {String} relative path to place test file (from PROJECT/tests)
+     * @param options.sampleTestGenerator.contents {String} contents of the test file the path that's returned
+     */
+    registerTestingFramework: function (name, options) {
+      _config[name] = _parseTestingFrameworkOptions(name, options);
+      // make sure the appropriate aggregate records are added
+      VelocityAggregateReports.insert({
+        name: name,
+        result: 'pending'
+      });
     }
+
   });
-
-  if (Meteor.isServer) {
-    _.extend(Velocity, {
-      /**
-       * Registers a testing framework plugin.
-       *
-       * @method registerTestingFramework
-       * @param {String} name                       The name of the testing framework.
-       * @param {Object} [options]                  Options for the testing framework.
-       * @param {String} options.disableAutoReset   Velocity's reset cycle will skip reports and logs for this framework
-       *                                            It will be the responsibility of the framework to clean up its ****!
-       * @param {String} options.regex              The regular expression for test files that should be assigned
-       *                                            to the testing framework.
-       *                                            The path relative to the tests
-       *                                            folder is matched against it.
-       *                                            The default is "name/.+\.js$"
-       *                                            (name is the testing framework name).
-       * @param options.sampleTestGenerator {Function} sampleTestGenerator
-       *    returns an array of fileObjects with the following fields:
-       * @param options.sampleTestGenerator.path {String} relative path to place test file (from PROJECT/tests)
-       * @param options.sampleTestGenerator.contents {String} contents of the test file the path that's returned
-       */
-      registerTestingFramework: function (name, options) {
-        _config[name] = _parseTestingFrameworkOptions(name, options);
-
-        // make sure the appropriate aggregate records are added
-        _reset(_config);
-      }
-    });
-  }
 
 
 //////////////////////////////////////////////////////////////////////
@@ -395,7 +371,7 @@ Velocity = {};
       });
 
       VelocityAggregateReports.upsert({'name': data.framework},
-                                      {$set: {'result': 'completed'}});
+        {$set: {'result': 'completed'}});
       _updateAggregateReports();
     },  // end completed
 
@@ -425,7 +401,7 @@ Velocity = {};
         sampleTests = _config[options.framework].sampleTestGenerator(options);
 
         DEBUG && console.log('[velocity] found ', sampleTests.length,
-                             'sample test files for', options.framework);
+          'sample test files for', options.framework);
 
         sampleTests.forEach(function (testFile) {
           var fullTestPath = path.join(Velocity.getTestsPath(), testFile.path);
@@ -437,20 +413,20 @@ Velocity = {};
       } else {
 
         samplesPath = path.join(Velocity.getAppPath(), 'packages',
-                                options.framework, 'sample-tests');
+          options.framework, 'sample-tests');
         testsPath = Velocity.getTestsPath();
 
         DEBUG && console.log('[velocity] checking for sample tests in',
-                              path.join(samplesPath, '*'));
+          path.join(samplesPath, '*'));
 
         if (fs.existsSync(samplesPath)) {
           command = 'mkdir -p ' + testsPath + ' && ' +
-                    'rsync -au ' + path.join(samplesPath, '*') +
-                    ' ' + testsPath + path.sep;
+          'rsync -au ' + path.join(samplesPath, '*') +
+          ' ' + testsPath + path.sep;
 
           DEBUG && console.log('[velocity] copying sample tests (if any) ' +
-                               'for framework', options.framework, '-',
-                               command);
+            'for framework', options.framework, '-',
+            command);
 
           child_process.exec(command, Meteor.bindEnvironment(
             function copySampleTestsExecHandler (err, stdout, stderr) {
@@ -465,93 +441,9 @@ Velocity = {};
         }
 
       }
-    },  // end copySampleTests
+    }  // end copySampleTests
 
 
-    /**
-     * Starts a new mirror if it has not already been started, and reuses an
-     * existing one if it is already started.
-     *
-     * This method will update the `VelocityMirrors` collection with `requestId`
-     * once the mirror is ready for use.
-     *
-     * @method velocity/mirrors/request
-     *
-     * @param {Object} options                  Options for the mirror.
-     * @param {String} options.framework        The name of the calling framework
-     * @param {String} [options.fixtureFiles]   Array of files with absolute paths
-     * @param {String} [options.port]           String use a specific port
-     * @param {String} [options.requestId]      Id for the mirror that may be used to query the mirror for status info
-     * @param {String} [options.rootUrlPath]    Adds this string to the end of the root url in the VelocityMirrors collection.
-     *                                          eg. `/?jasmine=true`
-     *                                          Request parameters that velocity-ci uses.
-     *
-     * @return {String} The `requestId` that can be used to query the mirror
-     */
-    'velocity/mirrors/request': function (options) {
-      check(options, {
-        framework: String,
-        port: Match.Optional(Number),
-        fixtureFiles: Match.Optional([String]),
-        requestId: Match.Optional(String),
-        rootUrlPath: Match.Optional(String)
-      });
-      options.port = options.port || 5000;
-      options.requestId = options.requestId || Random.id();
-
-      var mirrorUrl = _getMirrorUrl(options.port);
-      var rootUrlPath = (options.rootUrlPath || '').replace(/\//, '');
-      options.rootUrl = mirrorUrl + rootUrlPath;
-
-      DEBUG && console.log('[velocity] Mirror requested', options,
-                           'requestId:', options.requestId);
-
-
-      var connectToMirrorViaDDP = function () {
-        DEBUG && console.log('[velocity] Connecting to mirror via DDP...');
-        var ddpConnection = DDP.connect(mirrorUrl);
-        ddpConnection.onReconnect = function () {
-          DEBUG && console.log('[velocity] Connected, updating mirror metadata');
-          _updateMirrorMetadata(options);
-          this.disconnect();
-        };
-      };
-
-      DEBUG && console.log('[velocity] Checking if requested mirror is already started.');
-      if (!_isProcessRunningForRegisteredMirrorPid(options.port)) {
-        DEBUG && console.log('[velocity] Requested mirror not started. Starting...');
-        _velocityStartMirror(options, function () {
-          connectToMirrorViaDDP();
-        });
-      } else {
-        DEBUG && console.log('[velocity] Requested mirror pid already running. Reusing...');
-        connectToMirrorViaDDP();
-      }
-
-      return options.requestId;
-    },
-
-
-    /**
-     * Exposes the IS_MIRROR flag to clients
-     *
-     * @method velocity/isMirror
-     * @return {Boolean} true if currently running in mirror
-     */
-    'velocity/isMirror': function () {
-      return !!process.env.IS_MIRROR;
-    },
-
-    /**
-     * Syncs the mirror filesystem on an adhoc basis.
-     * Used by the core when file changes are detected.
-     *
-     * @method velocity/syncMirror
-     */
-    'velocity/syncMirror': function () {
-      DEBUG && console.log('[velocity] client restart requested velocity/syncMirror');
-      _syncMirror();
-    }
 
   });  // end Meteor methods
 
@@ -562,208 +454,13 @@ Velocity = {};
 // Private functions
 //
 
-
-  /**
-   * Starts a mirror and copies any specified fixture files into the mirror.
-   *
-   * @method velocityStartMirror
-   * @param {Object} options Required fields:
-   *                   framework - String ex. 'mocha-web-1'
-   *                   rootUrl - String ex. 'http://localhost:5000/x=y'
-   *
-   *                 Optional parameters:
-   *                   fixtureFiles - Array of files with absolute paths
-   *                   port - String use a specific port instead of finding the next available one
-   *                   next - function to call after the mirror has started
-   *
-   * @private
-   */
-  function _velocityStartMirror (options, next) {
-
-    var port = options.port;
-    var mongoLocation = _getMongoUrl(options.framework);
-    var mirrorLocation = _getMirrorUrl(port);
-
-    if (options.fixtureFiles) {
-      _addFixtures(options.fixtureFiles);
+  function _triggerVelocityStartupFunctions () {
+    _velocityStarted = true;
+    DEBUG && console.log('[velocity] Triggering queued startup functions');
+    var func;
+    while (func = _velocityStartupFunctions.pop()) {
+      func();
     }
-
-    var opts = {
-      cwd: Velocity.getMirrorPath(),
-      stdio: 'pipe',
-      env: _.extend({}, process.env, {
-        ROOT_URL: mirrorLocation,
-        MONGO_URL: mongoLocation,
-        PARENT_URL: process.env.ROOT_URL,
-        IS_MIRROR: true
-      })
-    };
-
-    var settingsPath = path.join(Velocity.getMirrorPath(), 'settings.json');
-    outputFile(settingsPath, JSON.stringify(Meteor.settings));
-
-    console.log('[velocity] Starting mirror at', mirrorLocation);
-
-
-    // perform a forced rsync as we are about to start a mirror
-    _syncMirror(true, function () {
-      _startMeteor(port, settingsPath, opts, function () {
-        // do another forced sync in case the user changes any files whilst the mirror is starting up
-        next();
-      });
-    });
-
-  } // end velocityStartMirror
-
-  /**
-   * Updated the VelocityMirrors collection with metadata about a newly started or reused mirror
-   *
-   * @method _updateMirrorMetadata
-   * @param {Object} options Required fields:
-   *                   framework - String ex. 'mocha-web-1'
-   *                   port - String the port the mirror that just started / was reused is using
-   *                   requestId - the request id to put in the mirror metadata
-   *
-   * @private
-   */
-  function _updateMirrorMetadata (options) {
-    // if this is a request we've seen before
-    var existingMirror = VelocityMirrors.findOne({
-      framework: options.framework,
-      port: options.port
-    });
-    if (existingMirror) {
-      // if we already have this mirror metadata, update it
-      VelocityMirrors.update(existingMirror._id, {$set: {requestId: options.requestId}});
-    } else {
-      // if this is a request we haven't seen before, create a new metadata entry
-      VelocityMirrors.upsert(
-        {
-          framework: options.framework,
-          port: options.port
-        }, {
-          framework: options.framework,
-          port: options.port,
-          rootUrl: options.rootUrl,
-          mongoUrl: _getMongoUrl(options.framework),
-          requestId: options.requestId
-        });
-    }
-  }
-
-  // TODO DOC IT UP
-  function _startMeteor (port, settingsPath, procOpts, next) {
-
-    var meteorArgs = [];
-    meteorArgs.push('--port', port);
-    meteorArgs.push('--settings', settingsPath);
-
-    var meteorProc = child_process.execFile('meteor', meteorArgs, procOpts);
-    meteorProc.stdout.pipe(process.stdout);
-    meteorProc.stderr.pipe(process.stderr);
-    var stopMeteorProc = function () {
-      DEBUG && console.log('[velocity] killing mirror');
-      meteorProc.kill();
-    };
-    process.on('SIGINT', stopMeteorProc);
-
-    meteorProc.on('exit', function (code, signal) {
-      DEBUG && console.log('[velocity] Mirror exited with code:', code, ' signal:', signal);
-      process.removeListener('SIGINT', stopMeteorProc);
-      // should we respawn here based on code/signal?
-    });
-
-    meteorProc.stdout.setEncoding('utf8');
-    var processMeteorStartup = Meteor.bindEnvironment(function (data) {
-      if (data.match(/Started Proxy/i)) {
-        console.log('[velocity] Mirror started.');
-        var mirrorPidId = MIRROR_PID_VAR_TEMPLATE.replace('@PORT', port);
-        VelocityVars.upsert({_id: mirrorPidId}, {_id: mirrorPidId, value: meteorProc.pid});
-        meteorProc.stdout.removeListener('data', processMeteorStartup);
-        next(null, meteorProc);
-      }
-      else if (data.match(/Perhaps another Meteor is running/i)) {
-        next(new Error(data));
-      }
-      // TODO there are more scenarios where meteor exists and we need to have those covered
-    });
-    meteorProc.stdout.on('data', processMeteorStartup);
-  }
-
-  /**
-   * Returns true if a mirror with the specified port was started by Velocity and if the pid stored
-   * at the time is currently running. This method DOES NOT check the actual connection
-   * @param port this is used to look up the PID for the mirror if it was started by Velocity
-   * @returns {string} true if the registered mirror process is running
-   * @private
-   */
-  function _isProcessRunningForRegisteredMirrorPid (port) {
-    var mirrorPid = MIRROR_PID_VAR_TEMPLATE.replace('@PORT', port);
-    var pid = VelocityVars.findOne(mirrorPid);
-    if (!pid) {
-      return false;
-    }
-    DEBUG && console.log('[velocity] Checking if mirror is running with pid', pid);
-    try {
-      process.kill(pid.value, 0);
-      DEBUG && console.log('[velocity] process with pid', pid, 'is running');
-      return true;
-    } catch (e) {
-      DEBUG && console.log('[velocity] process with ', pid, 'is not running');
-      return false;
-    }
-  }
-
-  function _getTestFrameworkNames () {
-    return _.pluck(_config, 'name');
-  }
-
-  /**
-   * Returns the MongoDB URL for the given database.
-   * @param database
-   * @returns {string} MongoDB Url
-   * @private
-   */
-  function _getMongoUrl (database) {
-    var mongoLocationParts = url.parse(process.env.MONGO_URL);
-    return url.format({
-      protocol: mongoLocationParts.protocol,
-      slashes: mongoLocationParts.slashes,
-      hostname: mongoLocationParts.hostname,
-      port: mongoLocationParts.port,
-      pathname: '/' + database
-    });
-  }
-
-  /**
-   * Return URL for the mirror with the given port.
-   * @param port Mirror port
-   * @returns {string} Mirror URL
-   * @private
-   */
-  function _getMirrorUrl (port) {
-    var rootUrlParts = url.parse(Meteor.absoluteUrl());
-    return url.format({
-      protocol: rootUrlParts.protocol,
-      slashes: rootUrlParts.slashes,
-      hostname: rootUrlParts.hostname,
-      port: port,
-      pathname: rootUrlParts.pathname
-    });
-  }
-
-  /**
-   * Add fixtures to the database.
-   * @param fixtureFiles Array with fixture file paths.
-   * @private
-   */
-  function _addFixtures (fixtureFiles) {
-    _.each(fixtureFiles, function (fixtureFile) {
-      VelocityFixtureFiles.insert({
-        _id: fixtureFile,
-        absolutePath: fixtureFile
-      });
-    });
   }
 
   function _parseTestingFrameworkOptions (name, options) {
@@ -783,27 +480,24 @@ Velocity = {};
    *
    * @method _initWatcher
    * @param {Object} config  See `registerTestingFramework`.
+   * @param {function} callback  Called after the watcher completes its first scan and is ready
    * @private
    */
-  function _initWatcher (config) {
+  function _initWatcher (config, callback) {
 
-    _watcher = chokidar.watch(Velocity.getTestsPath(), {ignored: /[\/\\]\./});
+    VelocityTestFiles.remove({});
+    VelocityFixtureFiles.remove({});
 
-    console.log('[velocity] chokadir watching ' + Velocity.getTestsPath());
+    _watcher = chokidar.watch(Velocity.getTestsPath(), {ignored: /[\/\\]\./, persistent: true});
 
     _watcher.on('add', Meteor.bindEnvironment(function (filePath) {
+
       var relativePath,
           targetFramework,
           data;
 
       filePath = path.normalize(filePath);
-
-      DEBUG && console.log('File added:', filePath);
-
-      relativePath = filePath.substring(Velocity.getAppPath().length);
-      if (relativePath[0] === path.sep) {
-        relativePath = relativePath.substring(1);
-      }
+      relativePath = _getRelativePath(filePath);
 
       // if this is a fixture file, put it in the fixtures collection
       if (FIXTURE_REG_EXP.test(relativePath)) {
@@ -811,10 +505,10 @@ Velocity = {};
         VelocityFixtureFiles.insert({
           _id: filePath,
           absolutePath: filePath,
+          relativePath: relativePath,
           lastModified: Date.now()
         });
-        // update the mirror right away
-        _syncMirror(true);
+        // bail early
         return;
       }
 
@@ -824,8 +518,9 @@ Velocity = {};
         return framework._regexp.test(relativePath);
       });
 
+
       if (targetFramework) {
-        DEBUG && console.log(targetFramework.name, ' <= ', filePath);
+        DEBUG && console.log('[velocity] Target framework for', relativePath, 'is', targetFramework.name);
 
         data = {
           _id: filePath,
@@ -836,28 +531,36 @@ Velocity = {};
           lastModified: Date.now()
         };
 
-        //DEBUG && console.log('data', data);
         VelocityTestFiles.insert(data);
+      } else {
+        DEBUG && console.log('[velocity] No framework registered for', relativePath);
       }
     }));  // end watcher.on 'add'
 
     _watcher.on('change', Meteor.bindEnvironment(function (filePath) {
-      DEBUG && console.log('File changed:', filePath);
+      DEBUG && console.log('[velocity] File changed:', _getRelativePath(filePath));
 
       // Since we key on filePath and we only add files we're interested in,
       // we don't have to worry about inadvertently updating records for files
       // we don't care about.
-      VelocityFixtureFiles.update(filePath, {$set: {lastModified: Date.now()}});
       VelocityTestFiles.update(filePath, {$set: {lastModified: Date.now()}});
     }));
 
     _watcher.on('unlink', Meteor.bindEnvironment(function (filePath) {
-      DEBUG && console.log('File removed:', filePath);
+      DEBUG && console.log('[velocity] File removed:', _getRelativePath(filePath));
       // If we only remove the file, we also need to remove the test results for
       // just that file. This required changing the postResult API and we could
       // do it, but the brute force method of reset() will do the trick until we
       // want to optimize VelocityTestFiles.remove(filePath);
+      VelocityTestFiles.remove({});
       _reset(config);
+    }));
+
+    _watcher.on('ready', Meteor.bindEnvironment(function () {
+      DEBUG && console.log('[velocity] File scan complete, now watching', Velocity.getTestsPath().substring(Velocity.getAppPath().length));
+      if (callback) {
+        callback();
+      }
     }));
 
   }  // end _initWatcher
@@ -873,18 +576,8 @@ Velocity = {};
 
     DEBUG && console.log('[velocity] resetting the world');
 
-    if (_watcher) {
-      _watcher.close();
-    }
-
-    VelocityTestFiles.remove({});
-    VelocityFixtureFiles.remove({});
-    VelocityFixtureFiles.insert({
-      _id: DEFAULT_FIXTURE_PATH,
-      absolutePath: DEFAULT_FIXTURE_PATH
-    });
     var frameworksWithDisableAutoReset = _(config).where({disableAutoReset: true}).pluck('name').value();
-    DEBUG && console.log('[velocity] not resetting reports and logs for', frameworksWithDisableAutoReset);
+    DEBUG && console.log('[velocity] frameworks with disable auto reset:', frameworksWithDisableAutoReset);
     VelocityTestReports.remove({framework: {$nin: frameworksWithDisableAutoReset}});
     VelocityLogs.remove({framework: {$nin: frameworksWithDisableAutoReset}});
     VelocityAggregateReports.remove({});
@@ -902,9 +595,7 @@ Velocity = {};
         result: 'pending'
       });
     });
-    VelocityMirrors.remove({});
 
-    _initWatcher(config);
   }
 
   /**
@@ -940,117 +631,24 @@ Velocity = {};
     if (aggregateComplete) {
       if ((aggregateComplete.result !== 'completed') && (_getTestFrameworkNames().length === completedFrameworksCount)) {
         VelocityAggregateReports.update({'name': 'aggregateComplete'}, {$set: {'result': 'completed'}});
-
-        _.each(_postProcessors, function (reporter) {
-          reporter();
+        _.each(Velocity.postProcessors, function (processor) {
+          processor();
         });
 
       }
     }
-
   }
 
-  /**
-   * Creates a physical mirror of the application under .meteor/local/.mirror
-   *
-   *     - Any files with the pattern tests/.*  are not copied, this stops .report
-   *     directory from also being copied.
-   *
-   *     TODO - Strips out velocity and reporters from the mirror's .meteor/packages file
-   *
-   * @method _syncMirror
-   * @param force performs an rsync even if no mirrors have been requested
-   * @private
-   */
-  var _syncing = false;
-
-  function _syncMirror (force, next) {
-
-    // de-bounce sync requests
-    if (_syncing) {
-      DEBUG && console.log('[velocity] de-bouncing sync mirror');
-      next && next();
-      return;
+  function _getRelativePath (filePath) {
+    var relativePath = filePath.substring(Velocity.getAppPath().length);
+    if (relativePath[0] === path.sep) {
+      relativePath = relativePath.substring(1);
     }
-
-    // don't sync if no mirrors have been requested, unless force is set
-    if (!force && VelocityMirrors.find().count() === 0) {
-      DEBUG && console.log('[velocity] not syncing as no mirror metadata was found.');
-      next && next();
-      return;
-    }
-
-    _syncing = true;
-    DEBUG && console.log('[velocity] syncing mirror');
-
-
-    var cmd = new Rsync()
-      .shell('ssh')
-      .flags('av')
-      .set('delete')
-      .set('delay-updates')
-      .exclude('/.meteor/local')
-      .exclude('/tests/.*')
-      .exclude('/packages')
-      // This file is created by velocity:core.
-      // It must be excluded to prevent rsync from deleting it.
-      .exclude('/settings.json')
-      // We must use escapeShellArgument until
-      // https://github.com/mattijs/node-rsync/pull/23
-      // has been pulled.
-      .source(escapeShellArgument(Velocity.getAppPath()) + path.sep)
-      .destination(escapeShellArgument(Velocity.getMirrorPath()));
-    var then = Date.now();
-    cmd.execute(Meteor.bindEnvironment(function (error) {
-
-      if (error) {
-        console.error('[velocity] Error syncing mirror', error);
-      } else {
-        DEBUG && console.log('[velocity] rsync took', Date.now() - then);
-      }
-
-      _copyFixtureFilesIntoMirror();
-      _symlinkPackagesDirIfPresent();
-
-      _.each(_preProcessors, function (preProcessor) {
-        preProcessor();
-      });
-
-      _syncing = false;
-
-      next && next();
-
-    }));
+    return relativePath;
   }
 
-  // DOC IT UP
-  function _copyFixtureFilesIntoMirror () {
-    DEBUG && console.log('[velocity] copying fixture files');
-    VelocityFixtureFiles.find({}).forEach(function (fixture) {
-      var fixtureLocationInMirror = Velocity.getMirrorPath() + path.sep + path.basename(fixture.absolutePath) + path.extname(fixture.absolutePath);
-      DEBUG && console.log('[velocity] adding fixture to watch list', fixture.absolutePath, 'to mirror');
-      copyFile(fixture.absolutePath, fixtureLocationInMirror);
-    });
-  }
-
-  /**
-   * Checks if the user has a local packages directory, if so it ensures it's symlinked in the mirror.
-   * The reason this is needed is because the standard rsync will copy a the packages dir and not respect
-   * the symlinks inside it.
-   *
-   * @method _symlinkPackagesDirIfPresent
-   * @private
-   */
-  function _symlinkPackagesDirIfPresent () {
-    var mirrorDir = Velocity.getMirrorPath(),
-        packagesDir = path.join(Velocity.getAppPath(), 'packages'),
-        mirrorPackagesDir = path.join(mirrorDir, 'packages');
-
-    if (fs.existsSync(packagesDir) && !fs.existsSync(mirrorPackagesDir)) {
-      DEBUG && console.log('[velocity] symlinking packages into mirror');
-      fs.symlinkSync(path.relative(mirrorDir, packagesDir), mirrorPackagesDir);
-    }
-
+  function _getTestFrameworkNames () {
+    return _.pluck(_config, 'name');
   }
 
 })();